scripts/changelog.sh: add special case for meta-switch rename
We recently renamed meta-switch to meta-bisdn-linux. This causes the
changelog script to think we removed meta-switch and added
meta-bisdn-linux, with no changelog between these two.

To get a proper changelog, add a special case to detect if we saw both
repositories and treat meta-switch as an older version of
meta-bisdn-linux.

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>