gh-workflow: prune sstate-cache
We need to keep the sum of the sstate-cache directories for both
MACHINES under 10 GiB (current github limitation for workflow cache). A
clean sstate-cache directory is about 4.8 GiB each.

We are bound to hit the limit sooner or later because the cache grows a)
when we add software to BISDN Linux and b) after each cache miss.

To address a), we keep cache entries only for the packages that benefit
the most (according to buildstats) from having their states cached. To
address b), we run sstate-cache-management.sh to remove duplicates.

Signed-off-by: Roger Luethi <roger.luethi@bisdn.de>